Added the currentTurn to the NumberCard to make sure the user cannot click a card while it is the Bot's turn

+++ src/app/modules/gameboard/components/number-card/number-card.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { BoardService } from '../../services/board.service';
 import {fadeInOut} from "../../animations/animations.animation";
+import { BoardMaster } from '../../interfaces/board-master.interface';
 @Component({
   selector: 'number-card',
   templateUrl: './number-card.component.html',
@@ -9,13 +10,19 @@ import {fadeInOut} from "../../animations/animations.animation";
 })
 export class NumberCardComponent implements OnInit {
   @Input() value!: number;
+  currentTurn!: string;
   constructor(private boardService : BoardService) { }
 
   ngOnInit(): void {
+    this.boardService.boardMaster.subscribe((boardMaster : BoardMaster) => {
+      this.currentTurn = boardMaster.currentTurn;
+    })
   }
 
   removeCard(): void {
-    this.boardService.removeCard(this.value);
+    if(this.currentTurn === "You") {
+      this.boardService.removeCard(this.value);
+    }
   }
 
 }
